chore: video 폴더 이동
