+++ apps/videos/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class VideoConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'video'
+    name = 'apps.videos'
