--- video/urls.py
@@ -1,10 +1,0 @@
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-from .views import VideoViewSet
-
-router = DefaultRouter()
-router.register(r'videos', VideoViewSet)
-
-urlpatterns = [
-    path('', include(router.urls)),
-]
